requirements.txt
streamlit
streamlit-drawable-canvas
numpy
matplotlib
pillow
tensorflow
